Extract FuzzySetShape/InitResult/InitMethod to sets/shape.py
