+++ src/fuzzy/sets/abstract.py
@@ -9,8 +9,6 @@ import inspect
 
 # import logging
 from abc import abstractmethod
-from dataclasses import dataclass
-from enum import Enum, auto
 from pathlib import Path
 from typing import Any, List, MutableMapping, NoReturn, Optional, Tuple, Type, Union
 
@@ -40,73 +38,9 @@ from ..utils.classes import DynamicParameterList, Loggable  # noqa: F401
 from .cache import MembershipCache, ParameterSignature, signature_of
 from .membership import Membership
 
-
-@dataclass(frozen=True)
-class FuzzySetShape:
-    """
-    A dataclass containing information about the shape of homogeneous fuzzy sets.
-    """
-
-    n_variables: int
-    n_terms: int
-
-
-@dataclass
-class FuzzySetInitResult:
-    """
-    A dataclass representing an initialization result concerning the centers and widths of fuzzy
-    set(s).
-    """
-
-    centers: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]]
-    widths: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]]
-
-
-class FuzzySetInitMethod(Enum):
-    """
-    An extended Enum that offers various built-in functionality for basic fuzzy set initialization
-    methods.
-    """
-
-    RANDOM = auto()
-    LINEAR = auto()
-
-    def initialize(
-        self,
-        shape: FuzzySetShape,
-        init_width: float = 1.0,
-    ) -> FuzzySetInitResult:
-        """
-        Perform a basic initialization of parameters for fuzzy sets.
-
-        Args:
-            shape: The shape of these fuzzy sets.
-            init_width: The initial width to use for each fuzzy set.
-
-        Returns:
-            An initialization result containing parameters, such as centers and widths.
-        """
-        if self is FuzzySetInitMethod.RANDOM:
-            centers: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]] = (
-                np.random.randn(shape.n_variables, shape.n_terms)
-            )
-            widths: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]] = np.abs(
-                np.random.randn(shape.n_variables, shape.n_terms)
-            ).clip(min=0.1)
-
-        elif self is FuzzySetInitMethod.LINEAR:
-            base = np.linspace(0.0, 1.0, num=shape.n_terms)
-            centers: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]] = (
-                np.repeat(base[None, :], repeats=shape.n_variables, axis=0)
-            )
-            widths: Union[float, ndarray[Any, np.dtype[np.floating[_64Bit]]]] = np.full(
-                (shape.n_variables, shape.n_terms), init_width, dtype=np.float32
-            )
-
-        else:
-            raise ValueError(f"Unsupported method: {self}")
-
-        return FuzzySetInitResult(centers, widths)
+# FuzzySetShape/FuzzySetInitResult/FuzzySetInitMethod re-exported here for backward
+# compatibility (they used to be defined in this module)
+from .shape import FuzzySetInitMethod, FuzzySetInitResult, FuzzySetShape  # noqa: F401
 
 
 class FuzzySet(TorchJitModule, Loggable, metaclass=abc.ABCMeta):
